+++ src/crypto/SymmetricCipherSalsa20.cpp
@@ -74,11 +74,11 @@ void SymmetricCipherSalsa20::processInPlace(QByteArray& data)
                          reinterpret_cast<u8*>(data.data()), data.size());
 }
 
-void SymmetricCipherSalsa20::processInPlace(QByteArray& data, int rounds)
+void SymmetricCipherSalsa20::processInPlace(QByteArray& data, quint64 rounds)
 {
     Q_ASSERT((data.size() < blockSize()) || ((data.size() % blockSize()) == 0));
 
-    for (int i = 0; i != rounds; ++i) {
+    for (quint64 i = 0; i != rounds; ++i) {
         ECRYPT_encrypt_bytes(&m_ctx, reinterpret_cast<const u8*>(data.constData()),
                              reinterpret_cast<u8*>(data.data()), data.size());
     }
